backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=2.3.0,<3.0.0",
    "flask-cors>=6.0.1",
    "langchain>=1.0.2",
    "langchain-openai>=0.2.0",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=0.2.0",
    "openai>=2.6.0",
    "pypdf2>=3.0.1",
    "flask-security-too>=5.3.3",
    "flask-login>=0.6.3",
    "mongoengine>=0.27.0",
    "email-validator>=2.1.0.post1",
    "setuptools>=65.0.0",
    "qdrant-client>=1.15.1",
    "langgraph-checkpoint-mongodb>=0.3.0",
]
